backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
email-validator
requests
